refactor(worker): give the model wiring its own file, like cmd/api has
cmd/api keeps its model-path resolution in modelwiring.go; cmd/worker
kept the same three declarations at the bottom of main.go, which the
500-line file cap now refuses. Move them to the sibling file of the
same name — the file each role's boot reads to answer "what serves the
AI surfaces here".
